Reformatted code
Proper use of controller and helper.
